add support for get/set hidden files and directories. DRY up some windows native code.

+++ src/windows/fs_meta.h
@@ -2,27 +2,27 @@
 #pragma once
 #include "../common/volume_metadata.h"
 #include "../common/volume_mount_points.h"
-#include "./error_utils.h"
 #include <napi.h>
-
-#include <string>
+#include <windows.h> // for MAX_PATH
+#include <functional>
+#include <utility>
 
 namespace FSMeta {
 
 constexpr size_t ERROR_BUFFER_SIZE = 256;
-constexpr DWORD BUFFER_SIZE = MAX_PATH + 1;
+constexpr size_t BUFFER_SIZE = MAX_PATH + 1;
 
 enum DriveStatus {
-  Unknown,
-  Unavailable,
-  Healthy,
-  Disconnected,
-  Error,
-  NoMedia
+    Unknown,
+    Unavailable, 
+    Healthy,
+    Disconnected,
+    Error,
+    NoMedia
 };
 
 inline const char *DriveStatusToString(DriveStatus status) {
-  switch (status) {
+    switch (status) {
   case Unknown:
     return "unknown";
   case Unavailable:
@@ -37,7 +37,7 @@ inline const char *DriveStatusToString(DriveStatus status) {
     return "no_media";
   default:
     return "unknown";
-  }
+    }
 }
 
 } // namespace FSMeta
